RF: moved webui related ls code into ls_webui (the same for tests)
To get some breething room in ls
